DCMPG should pop doubles and not floats; Add MLEAudioStream for SpringCoat; Correct MLEMathAccel methods for SpringCoat having the wrong return types.

+++ emulators/springcoat-vm/src/main/java/cc/squirreljme/vm/springcoat/SpringThreadWorker.java
@@ -2059,8 +2059,8 @@ public final class SpringThreadWorker
 					// Compare double, NaN is positive
 				case InstructionIndex.DCMPG:
 					{
-						double b = frame.<Float>popFromStack(Float.class),
-							a = frame.<Float>popFromStack(Float.class);
+						double b = frame.<Double>popFromStack(Double.class),
+							a = frame.<Double>popFromStack(Double.class);
 						
 						if (Double.isNaN(a) || Double.isNaN(b))
 							frame.pushToStack(1);
